Move pipeline-building helpers to pipeline/builder subpackage
buildPipeline, mergeConnectorDefaults, and validatePipeline lived in
cmd/coordinator/main.go only because main was the one package able to
depend on both pipeline and steps. They cannot move into pipeline
directly: steps imports pipeline, so pipeline importing steps would
form a cycle, and pipeline.Build already names the per-step registry
builder. A pipeline/builder subpackage depends on both without a cycle,
leaving main as pure process wiring.

Pure reorganization; no behavior change. TestValidatePipeline moves with
the code.
