+++ cmd/coordinator/main.go
@@ -19,7 +19,6 @@ package main
 import (
 	"context"
 	"errors"
-	"fmt"
 	"net/http"
 	"os"
 	"os/signal"
@@ -34,8 +33,8 @@ import (
 	"github.com/llm-d/llm-d-router/pkg/coordinator/config"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/gateway"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/pipeline"
+	"github.com/llm-d/llm-d-router/pkg/coordinator/pipeline/builder"
 	"github.com/llm-d/llm-d-router/pkg/coordinator/server"
-	"github.com/llm-d/llm-d-router/pkg/coordinator/steps"
 )
 
 func main() {
@@ -82,7 +81,7 @@ func main() {
 
 	gwClient := gateway.New(cfg.Gateway)
 
-	steps, err := buildPipeline(cfg, gwClient)
+	steps, err := builder.Build(cfg, gwClient)
 	if err != nil {
 		log.Error(err, "failed to build pipeline")
 		os.Exit(1)
@@ -127,54 +126,4 @@ func serveUntilSignal(srv *server.Server, shutdownTimeout time.Duration) error {
 		}
 	}
 	return nil
-}
-
-// validatePipeline rejects configurations that cannot work before any step runs.
-// The tokens-in format (use_openai_format=false) sends token IDs that only the
-// render step produces, so it requires a render step in the pipeline.
-func validatePipeline(p config.PipelineConfig) error {
-	if p.UseOpenAIFormat {
-		return nil
-	}
-	for _, s := range p.Steps {
-		if s.Type == steps.RenderStepName {
-			return nil
-		}
-	}
-	return fmt.Errorf("pipeline.use_openai_format=false requires a %q step (the tokens-in format sends token IDs that render produces)", steps.RenderStepName)
-}
-
-func mergeConnectorDefaults(params map[string]any, kvConnector, ecConnector string) map[string]any {
-	out := make(map[string]any, len(params))
-	for k, v := range params {
-		out[k] = v
-	}
-	if _, ok := out[steps.ParamKVConnector]; !ok && kvConnector != "" {
-		out[steps.ParamKVConnector] = kvConnector
-	}
-	if _, ok := out[steps.ParamECConnector]; !ok && ecConnector != "" {
-		out[steps.ParamECConnector] = ecConnector
-	}
-	return out
-}
-
-func buildPipeline(cfg *config.Config, gwClient *gateway.Client) ([]pipeline.Step, error) {
-	if err := validatePipeline(cfg.Pipeline); err != nil {
-		return nil, err
-	}
-
-	var steps []pipeline.Step
-	for _, stepCfg := range cfg.Pipeline.Steps {
-		params := mergeConnectorDefaults(stepCfg.Params, cfg.Pipeline.KVConnector, cfg.Pipeline.ECConnector)
-		if _, ok := params["use_openai_format"]; !ok {
-			params["use_openai_format"] = cfg.Pipeline.UseOpenAIFormat
-		}
-		step, err := pipeline.Build(stepCfg.Type, gwClient, params)
-		if err != nil {
-			return nil, err
-		}
-
-		steps = append(steps, step)
-	}
-	return steps, nil
 }
